Modified NUMERIC_ARRAY type to map to array of integers.

+++ src/main/java/uk/ac/ebi/hca/importer/excel/NodeNavigator.java
@@ -83,7 +83,7 @@ class NodeNavigator {
         ((ObjectNode) currentNode).put(nextProperty, value);
     }
 
-    public void putNext(double[] value) {
+    public void putNext(int[] value) {
         ArrayNode array = ((ObjectNode) currentNode).putArray(nextProperty);
         Arrays.stream(value).forEach(array::add);
     }
